# packages/service/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/members/check-user-id": {
      "get": {
        "description": "회원가입 전에 userId가 이미 사용 중인지 확인합니다. available=true이면 가입에 사용할 수 있습니다.",
        "operationId": "MemberController_checkUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "description": "중복 여부를 확인할 회원 로그인 ID. 소문자로 시작하고 소문자/숫자/_ 조합 4~20자",
            "schema": {
              "pattern": "^[a-z][a-z0-9_]{3,19}$",
              "example": "movie_user",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "회원 아이디 사용 가능 여부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUserIdAvailabilityResultDto"
                }
              }
            }
          },
          "400": {
            "description": "userId 형식이 유효하지 않은 경우"
          }
        },
        "summary": "회원 아이디 중복 검사",
        "tags": [
          "Members"
        ]
      }
    },
    "/phone-verifications": {
      "post": {
        "description": "휴대전화번호 인증을 시작합니다. 실제 SMS 발송 대신 개발용 인증 코드를 응답으로 반환합니다.",
        "operationId": "MemberController_requestPhoneVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPhoneVerificationRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "휴대전화 인증 코드 발급 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneVerificationIssuedDto"
                }
              }
            }
          },
          "400": {
            "description": "휴대전화번호 형식이 유효하지 않은 경우"
          }
        },
        "summary": "휴대전화 인증 코드 발급",
        "tags": [
          "Members"
        ]
      }
    },
    "/phone-verifications/confirm": {
      "post": {
        "description": "발급된 인증 코드와 휴대전화번호를 검증하고 인증을 완료합니다.",
        "operationId": "MemberController_confirmPhoneVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPhoneVerificationRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "휴대전화 인증 확인 결과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneVerificationConfirmedDto"
                }
              }
            }
          },
          "400": {
            "description": "인증 요청이 없거나, 만료되었거나, 코드가 일치하지 않는 경우"
          }
        },
        "summary": "휴대전화 인증 코드 확인",
        "tags": [
          "Members"
        ]
      }
    },
    "/members/signup": {
      "post": {
        "description": "아이디 중복, 휴대전화 인증 완료 여부를 검증한 뒤 회원을 생성합니다.",
        "operationId": "MemberController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupMemberRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupMemberResultDto"
                }
              }
            }
          },
          "400": {
            "description": "입력값이 유효하지 않거나 휴대전화 인증이 완료되지 않은 경우"
          },
          "409": {
            "description": "회원 아이디 또는 휴대전화번호가 이미 사용 중인 경우"
          }
        },
        "summary": "회원가입",
        "tags": [
          "Members"
        ]
      }
    },
    "/members/login": {
      "post": {
        "description": "회원 아이디와 비밀번호를 검증합니다. 비밀번호 실패가 5회 누적되면 회원이 잠금 상태가 됩니다.",
        "operationId": "MemberController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginMemberRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginMemberResultDto"
                }
              }
            }
          },
          "400": {
            "description": "아이디/비밀번호가 일치하지 않거나 입력값이 유효하지 않은 경우"
          },
          "403": {
            "description": "비밀번호 실패 5회 누적으로 잠긴 회원인 경우"
          },
          "404": {
            "description": "회원을 찾을 수 없는 경우"
          }
        },
        "summary": "로그인",
        "tags": [
          "Members"
        ]
      }
    },
    "/members/temporary-password": {
      "post": {
        "description": "휴대전화 인증이 완료된 회원에게 임시비밀번호를 발급하고 잠금 상태를 해제합니다.",
        "operationId": "MemberController_issueTemporaryPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueTemporaryPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "임시비밀번호 발급 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemporaryPasswordIssuedDto"
                }
              }
            }
          },
          "400": {
            "description": "휴대전화 인증이 완료되지 않았거나 입력값이 유효하지 않은 경우"
          },
          "404": {
            "description": "회원을 찾을 수 없는 경우"
          }
        },
        "summary": "임시비밀번호 발급",
        "tags": [
          "Members"
        ]
      }
    },
    "/members/password": {
      "post": {
        "description": "기존 비밀번호 검증이 통과하면 신규 비밀번호로 변경합니다.",
        "operationId": "MemberController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMemberPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "비밀번호 변경 완료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPasswordChangedDto"
                }
              }
            }
          },
          "400": {
            "description": "기존 비밀번호가 일치하지 않거나 입력값이 유효하지 않은 경우"
          },
          "404": {
            "description": "회원을 찾을 수 없는 경우"
          }
        },
        "summary": "비밀번호 변경",
        "tags": [
          "Members"
        ]
      }
    },
...
  },
  "info": {
    "title": "GC Movie Reservation Service API",
    "description": "영화 예매 서비스 백엔드 API 문서입니다.\n회원가입/로그인, 휴대전화 인증, 주소 검색, 영화 목록, 영화관 목록 API를 제공합니다.\nValidationPipe가 whitelist와 forbidNonWhitelisted를 사용하므로 문서에 없는 요청 필드는 거부됩니다.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CheckUserIdAvailabilityResultDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean",
            "example": true,
            "description": "회원 아이디 사용 가능 여부"
          }
        },
        "required": [
          "available"
        ]
      },
      "RequestPhoneVerificationRequestDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "example": "01012345678",
            "pattern": "^\\d{10,11}$",
            "description": "인증 코드를 발급할 휴대전화번호"
          }
        },
        "required": [
          "phoneNumber"
        ]
      },
      "PhoneVerificationIssuedDto": {
        "type": "object",
        "properties": {
          "verificationId": {
            "type": "string",
            "example": "1",
            "description": "휴대전화 인증 요청 ID"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "개발 환경에서 클라이언트가 확인할 인증 코드"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-04-28T01:05:00.000Z",
            "description": "인증 코드 만료 시각"
          }
        },
        "required": [
          "verificationId",
          "code",
          "expiresAt"
        ]
      },
      "ConfirmPhoneVerificationRequestDto": {
        "type": "object",
        "properties": {
          "verificationId": {
            "type": "string",
            "example": "1",
            "description": "휴대전화 인증 요청 ID"
          },
          "phoneNumber": {
            "type": "string",
            "example": "01012345678",
            "pattern": "^\\d{10,11}$",
            "description": "인증 대상 휴대전화번호"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "pattern": "^\\d{6}$",
            "description": "발급받은 6자리 인증 코드"
          }
        },
        "required": [
          "verificationId",
          "phoneNumber",
          "code"
        ]
      },
      "PhoneVerificationConfirmedDto": {
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean",
            "example": true,
            "description": "휴대전화 인증 완료 여부"
          }
        },
        "required": [
          "verified"
        ]
      },
      "SignupMemberRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "movie_user",
            "pattern": "^[a-z][a-z0-9_]{3,19}$",
            "description": "회원 로그인 ID"
          },
          "password": {
            "type": "string",
            "example": "password123!",
            "minLength": 8,
            "maxLength": 72,
            "description": "회원 비밀번호"
          },
          "name": {
            "type": "string",
            "example": "홍길동",
            "minLength": 1,
            "maxLength": 50,
            "description": "회원 이름"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-01-01",
            "description": "생년월일. ISO date 문자열"
          },
          "phoneNumber": {
            "type": "string",
            "example": "01012345678",
            "pattern": "^\\d{10,11}$",
            "description": "휴대전화번호"
          },
          "address": {
            "type": "string",
            "example": "서울특별시 강남구 테헤란로 427",
            "minLength": 1,
            "maxLength": 255,
            "description": "회원 주소"
          },
          "phoneVerificationId": {
            "type": "string",
            "example": "1",
            "description": "인증 완료된 휴대전화 인증 요청 ID"
          }
        },
        "required": [
          "userId",
          "password",
          "name",
          "birthDate",
          "phoneNumber",
          "address",
          "phoneVerificationId"
        ]
      },
      "SignupMemberResultDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "example": "1",
            "description": "가입된 회원 PK"
          },
          "userId": {
            "type": "string",
            "example": "movie_user",
            "description": "회원 로그인 ID"
          }
        },
        "required": [
          "memberId",
          "userId"
        ]
      },
      "LoginMemberRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "movie_user",
            "pattern": "^[a-z][a-z0-9_]{3,19}$",
            "description": "회원 로그인 ID"
          },
          "password": {
            "type": "string",
            "example": "password123!",
            "minLength": 8,
            "maxLength": 72,
            "description": "회원 비밀번호"
          }
        },
        "required": [
          "userId",
          "password"
        ]
      },
      "LoginMemberResultDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "example": "1",
            "description": "로그인된 회원 PK"
          },
          "userId": {
            "type": "string",
            "example": "movie_user",
            "description": "회원 로그인 ID"
          }
        },
        "required": [
          "memberId",
          "userId"
        ]
      },
      "IssueTemporaryPasswordRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "movie_user",
            "pattern": "^[a-z][a-z0-9_]{3,19}$",
            "description": "임시비밀번호를 발급받을 회원 로그인 ID"
          },
          "phoneVerificationId": {
            "type": "string",
            "example": "1",
            "description": "인증 완료된 휴대전화 인증 요청 ID"
          }
        },
        "required": [
          "userId",
          "phoneVerificationId"
        ]
      },
      "TemporaryPasswordIssuedDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "movie_user",
            "description": "임시비밀번호가 발급된 회원 로그인 ID"
          },
          "temporaryPassword": {
            "type": "string",
            "example": "Temp-abc1231!",
            "description": "휴대전화 인증 후 발급된 임시비밀번호"
          }
        },
        "required": [
          "userId",
          "temporaryPassword"
        ]
      },
      "ChangeMemberPasswordRequestDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "movie_user",
            "pattern": "^[a-z][a-z0-9_]{3,19}$",
            "description": "회원 로그인 ID"
          },
          "currentPassword": {
            "type": "string",
            "example": "password123!",
            "minLength": 8,
            "maxLength": 72,
            "description": "기존 비밀번호"
          },
          "newPassword": {
            "type": "string",
            "example": "newPassword123!",
            "minLength": 8,
            "maxLength": 72,
            "description": "신규 비밀번호"
          }
        },
        "required": [
          "userId",
          "currentPassword",
          "newPassword"
        ]
      },
      "MemberPasswordChangedDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "movie_user",
            "description": "비밀번호가 변경된 회원 로그인 ID"
          },
          "changed": {
            "type": "boolean",
            "example": true,
            "description": "비밀번호 변경 완료 여부"
          }
        },
        "required": [
          "userId",
          "changed"
        ]
      },
...
    }
  }
}